% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coronavirus_region.R
\docType{data}
\name{coronavirus_region}
\alias{coronavirus_region}
\title{Australia coronavirus data}
\format{
a tibble with 2120 observations and 8 variables:
\itemize{
\item \strong{Country}: full country name
\item \strong{State}: one of Australian Capital Territory, New South Wales, Northern Territory, Queensland, South Australia, Tasmania, Victoria, Western Australia
\item \strong{date}: date from 2020-01-12 to 2020-10-02
\item \strong{confirmed}: the cumulative number of the confirmed cases
\item \strong{recovered}: the cumulative number of the recovered cases
\item \strong{deaths}: the cumulative number of the death cases
\item \strong{latitude}: the latitude of the state
\item \strong{longitude}: the longitude of the state
}
}
\source{
\href{https://covid19datahub.io/articles/data.html}{COVID-19 Data Hub}
}
\usage{
coronavirus_region
}
\description{
This dataset contains coronavirus data of 3 different types from 8 states in Australia.
}
\keyword{datasets}
