% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uifunction.R
\name{dateselection}
\alias{dateselection}
\title{Create a date range input}
\usage{
dateselection(starttime, endtime)
}
\arguments{
\item{starttime}{the beginning of the date}

\item{endtime}{the end of the date}
}
\value{
It shows a date range input in shiny app and related interactive plots and table.

If the input in user interface does not satisfy the condition,the shiny app will not display the plots and table.
}
\description{
Create a date range input
}
