% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch a shiny app about Australia coronavirus}
\usage{
launch_app()
}
\value{
A shiny app contains case data about Australia coronavirus
}
\description{
Launch a shiny app about Australia coronavirus
}
