% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{n_states}
\alias{n_states}
\title{Count the number of states in a dataset}
\usage{
n_states(data)
}
\arguments{
\item{data}{A dataset containing a \code{State} column.}
}
\description{
Count the number of states in a dataset
}
