% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectInputstate.R
\name{selectInputstate}
\alias{selectInputstate}
\title{Get the different states in Australia}
\usage{
selectInputstate(id)
}
\value{
The name list of different states
}
\description{
Get the different states in Australia
}
